First version of CCS support in the DCS Gateway.

+++ trunk/Repeater/DCSGateway/DCSGatewayThread.cpp
@@ -38,6 +38,7 @@ m_reflectors(),
 m_hostnames(),
 m_repeaterHandler(NULL),
 m_dcsHandler(),
+m_ccsHandler(),
 m_audioUnit(NULL),
 m_echoUnit(NULL),
 m_header(),
@@ -56,6 +57,7 @@ m_reconnect(RECONNECT_NEVER),
 m_language(TL_ENGLISH_UK),
 m_reconnectTimer(1000U, 0U),
 m_watchdogTimer(1000U, 2U),
+m_ccsTimer(1000U, 300U),
 m_id(0U),
 m_seq(0U),
 m_echo(false),
@@ -114,6 +116,8 @@ void* CDCSGatewayThread::Entry()
 		link(m_startupReflector);
 	}
 
+	m_ccsHandler.registration(CCS_HOST, CCS_PORT);
+
 	wxStopWatch timer;
 
 	while (!m_killed) {
@@ -124,7 +128,8 @@ void* CDCSGatewayThread::Entry()
 			CDCSGatewayAMBEData* data;
 
 			bool rptData = true;
-			bool deData  = true;
+			bool dcsData = true;
+			bool ccsData = true;
 
 			REPEATER_TYPE rptType = m_repeaterHandler->read();
 			switch (rptType) {
@@ -138,6 +143,7 @@ void* CDCSGatewayThread::Entry()
 						wxLogMessage(wxT("Repeater header - My: %s/%s  Your: %s  Rpt1: %s  Rpt2: %s  Flags: %02X %02X %02X"), header->getMyCall1().c_str(), header->getMyCall2().c_str(), header->getYourCall().c_str(), header->getRptCall1().c_str(), header->getRptCall2().c_str(), header->getFlag1(), header->getFlag2(), header->getFlag3());
 						showHeader(*header);
 						commandHandler(header->getYourCall());
+						m_ccsHandler.writeUserStatus(*header);
 						if (m_echo)
 							m_echoUnit->writeHeader(*header);
 						delete header;
@@ -150,24 +156,33 @@ void* CDCSGatewayThread::Entry()
 				case RT_DATA:
 					data = m_repeaterHandler->readData();
 					if (data != NULL) {
-						if (m_state == LINK_LINKED && !m_echo && !m_info) {
-							// Get the slow data text
-							if (m_text.IsEmpty()) {
-								m_textCollector.writeData(*data);
+						// Get the slow data text
+						if (m_text.IsEmpty()) {
+							m_textCollector.writeData(*data);
 
-								bool hasData = m_textCollector.hasData();
-								if (hasData) {
-									m_text = m_textCollector.getData();
-									CDCSGatewayAMBEData::setText(m_text);
-								}
+							bool hasData = m_textCollector.hasData();
+							if (hasData) {
+								m_text = m_textCollector.getData();
+								CDCSGatewayAMBEData::setText(m_text);
 							}
+						}
 
-							data->setRptSeq(m_seq);
-							m_seq++;
-
-							m_dcsHandler.writeData(m_header, *data);
-						} else if (m_echo) {
+						if (m_echo) {
 							m_echoUnit->writeData(*data);
+						} else if (m_info) {
+							// Do nothing
+						} else {
+							if (m_state == LINK_LINKED && (m_yourCall.Right(1U).IsSameAs(wxT("L")) || m_yourCall.Left(4U).IsSameAs(wxT("CQCQ")))) {
+								data->setRptSeq(m_seq);
+								m_seq++;
+
+								m_dcsHandler.writeData(m_header, *data);
+							} else if (m_state == LINK_CCS) {
+								data->setRptSeq(m_seq);
+								m_seq++;
+
+								m_ccsHandler.writeData(m_header, *data);
+							}
 						}
 
 						bool end = data->isEnd();
@@ -187,7 +202,7 @@ void* CDCSGatewayThread::Entry()
 			DCS_TYPE dcsType = m_dcsHandler.read();
 			switch (dcsType) {
 				case DT_NONE:
-					deData = false;
+					dcsData = false;
 					break;
 				case DT_DATA:
 					data = m_dcsHandler.readData();
@@ -230,12 +245,77 @@ void* CDCSGatewayThread::Entry()
 					break;
 			}
 
+			CCS_TYPE ccsType = m_ccsHandler.read();
+			switch (ccsType) {
+				case CT_NONE:
+					ccsData = false;
+					break;
+				case CT_DATA:
+					data = m_ccsHandler.readData();
+					if (data != NULL) {
+						unsigned int id = data->getId();
+						if (m_id == 0x00U) {
+							CDCSGatewayHeaderData* header = m_ccsHandler.getHeader();
+							if (header != NULL) {
+								wxLogMessage(wxT("CCS header - My: %s/%s  Your: %s"), header->getMyCall1().c_str(), header->getMyCall2().c_str(), header->getYourCall().c_str());
+								m_id = id;
+								showHeader(*header);
+								writeHeader(*header);
+								m_watchdogTimer.start();
+							}
+							delete header;
+						}
+
+						if (id == m_id) {
+							unsigned int seq = data->getRptSeq();
+
+							// Repeat the header every two seconds
+							if ((seq % 100U) == 0U) {
+								CDCSGatewayHeaderData* header = m_ccsHandler.getHeader();
+								writeHeader(*header);
+								delete header;
+							}
+
+							writeData(*data);
+							m_watchdogTimer.reset();
+
+							bool end = data->isEnd();
+							if (end) {
+								m_watchdogTimer.stop();
+								clearHeader();
+								m_id = 0x00U;
+							}
+						}
+						delete data;
+					}
+					break;
+				case CT_TERMINATE:
+					if (m_state == LINK_CCS) {
+						wxLogMessage(wxT("CCS routing terminated by remote system"));
+						m_ccsTimer.stop();
+						m_state = LINK_NONE;
+					}
+					ccsData = false;
+					break;
+				case CT_QUERY:
+					ccsData = false;
+					m_ccsHandler.writeQueryReply(m_state == LINK_CCS, m_yourCall, m_ccsTimer.getRemaining());
+					break;
+				case CT_QUERY_REPLY:
+					ccsData = false;
+					break;
+			}
+
 			// Any activity resets the reconnect timer
-			if (rptData || deData)
+			if (rptData || dcsData || ccsData) {
+				if (m_state == LINK_CCS)
+					m_ccsTimer.start();
+
 				m_reconnectTimer.start();
+			}
 
 			// No more work to be done?
-			if (!rptData && !deData)
+			if (!rptData && !dcsData && !ccsData)
 				break;
 		}
 
@@ -244,6 +324,12 @@ void* CDCSGatewayThread::Entry()
 			clearHeader();
 			m_watchdogTimer.stop();
 			m_id = 0x00U;
+		}
+
+		if (m_ccsTimer.isRunning() && m_ccsTimer.hasExpired()) {
+			wxLogMessage(wxT("CCS timer has expired"));
+			m_ccsHandler.writeTerminate();
+			m_ccsTimer.stop();
 		}
 
 		if (m_reconnectTimer.isRunning() && m_reconnectTimer.hasExpired()) {
@@ -256,7 +342,7 @@ void* CDCSGatewayThread::Entry()
 			}
 		}
 
-		if (m_state != LINK_NONE) {
+		if (m_state == LINK_LINKING || m_state == LINK_LINKED) {
 			bool connected = m_dcsHandler.isConnected();
 
 			if (m_state == LINK_LINKING && connected) {
@@ -282,6 +368,7 @@ void* CDCSGatewayThread::Entry()
 		m_dcsHandler.clock(ms);
 		m_reconnectTimer.clock(ms);
 		m_watchdogTimer.clock(ms);
+		m_ccsTimer.clock(ms);
 
 		m_audioUnit->clock(ms);
 		m_echoUnit->clock(ms);
@@ -290,6 +377,7 @@ void* CDCSGatewayThread::Entry()
 	wxLogMessage(wxT("Stopping the DCS Gateway thread"));
 
 	m_dcsHandler.unlink();
+	m_ccsHandler.cancelation();
 
 	m_repeaterHandler->close();
 	delete m_repeaterHandler;
@@ -308,9 +396,12 @@ void CDCSGatewayThread::kill()
 	m_killed = true;
 }
 
-void CDCSGatewayThread::setReflector(const wxString& callsign, const wxString& reflector, bool atStartup, RECONNECT reconnect, TEXT_LANG language)
+void CDCSGatewayThread::setReflector(const wxString& callsign, const wxString& locator, const wxString& reflector, bool atStartup, RECONNECT reconnect, TEXT_LANG language)
 {
 	m_dcsHandler.setCallsign(callsign);
+	m_ccsHandler.setCallsign(callsign);
+
+	m_ccsHandler.setLocator(locator);
 
 	m_startupReflector = reflector;
 	m_atStartup        = atStartup;
@@ -374,56 +465,83 @@ void CDCSGatewayThread::commandHandler(const wxString& yourCall)
 		return;
 	}
 
+	if (yourCall.IsSameAs(wxT("       X"))) {
+		if (m_state == LINK_CCS) {
+			wxLogMessage(wxT("CCS routing ended by %s"), m_myCall.c_str());
+			m_ccsHandler.writeTerminate();
+			m_ccsTimer.stop();
+			m_state = LINK_NONE;
+		}
+		return;
+	}
+
 	if (m_reconnect == RECONNECT_FIXED)
 		return;
 
-	wxString letter = yourCall.Right(1);
+	wxString letter = yourCall.Right(1U);
 
 	if (letter.IsSameAs(wxT("U"))) {
-		writeNotLinked();
+		if (m_state == LINK_LINKING || m_state == LINK_LINKED) {
+			writeNotLinked();
 
-		wxLogMessage(wxT("Unlink command issued by %s"), m_myCall.c_str());
-		unlink();
+			wxLogMessage(wxT("Unlink command issued by %s"), m_myCall.c_str());
+			unlink();
+		}
 	} else if (letter.IsSameAs(wxT("L"))) {
-		// Extract the callsign "1234567L" -> "123456 7"
-		wxString reflector = yourCall.Left(LONG_CALLSIGN_LENGTH - 2U);
-		reflector.Append(wxT(" "));
-		reflector.Append(yourCall.Mid(LONG_CALLSIGN_LENGTH - 2U, 1));
+		if (m_state != LINK_CCS) {
+			// Extract the callsign "1234567L" -> "123456 7"
+			wxString reflector = yourCall.Left(LONG_CALLSIGN_LENGTH - 2U);
+			reflector.Append(wxT(" "));
+			reflector.Append(yourCall.Mid(LONG_CALLSIGN_LENGTH - 2U, 1));
 
-		// Check for a duplicate link command
-		if (m_reflector.IsSameAs(reflector))
-			return;
+			// Check for a duplicate link command
+			if (m_reflector.IsSameAs(reflector))
+				return;
 
-		// Validate the reflector callsign
-		wxString baseName = reflector.Left(LONG_CALLSIGN_LENGTH - 1U).Trim();
-		int index = m_reflectors.Index(baseName);
-		if (index == wxNOT_FOUND) {
-			wxLogMessage(wxT("Invalid link command to %s issued by %s"), reflector.c_str(), m_myCall.c_str());
-			return;
+			// Validate the reflector callsign
+			wxString baseName = reflector.Left(LONG_CALLSIGN_LENGTH - 1U).Trim();
+			int index = m_reflectors.Index(baseName);
+			if (index == wxNOT_FOUND) {
+				wxLogMessage(wxT("Invalid link command to %s issued by %s"), reflector.c_str(), m_myCall.c_str());
+				return;
+			}
+
+			wxLogMessage(wxT("Link command to %s issued by %s"), reflector.c_str(), m_myCall.c_str());
+			unlink();
+			link(reflector);
+
+			writeLinkingTo(reflector);
+		}
+	} else if (!yourCall.Left(4U).IsSameAs(wxT("CQCQ"))) {
+		if (m_state != LINK_CCS) {
+			wxLogMessage(wxT("CCS route started by %s -> %s"), m_myCall.c_str(), yourCall.c_str());
+
+			writeNotLinked();
+			unlink();
+
+			m_state = LINK_CCS;
 		}
 
-		wxLogMessage(wxT("Link command to %s issued by %s"), reflector.c_str(), m_myCall.c_str());
-		unlink();
-		link(reflector);
-
-		writeLinkingTo(reflector);
+		m_ccsTimer.start();
+		m_ccsHandler.setRemote(yourCall);
 	}
 }
 
 void CDCSGatewayThread::unlink()
 {
 	// Anything to do?
-	if (m_state == LINK_NONE)
+	if (m_state != LINK_LINKING && m_state != LINK_LINKED)
 		return;
 
 	m_dcsHandler.unlink();
+	m_ccsHandler.setReflector(wxEmptyString);
 
 	m_state = LINK_NONE;
 }
 
 bool CDCSGatewayThread::link(const wxString &reflector)
 {
-	if (m_state != LINK_NONE)
+	if (m_state == LINK_LINKING || m_state == LINK_LINKED)
 		unlink();
 
 	wxString baseName = reflector.Left(LONG_CALLSIGN_LENGTH - 1U).Trim();
@@ -442,6 +560,7 @@ bool CDCSGatewayThread::link(const wxString &reflector)
 		m_state     = LINK_LINKING;
 		m_header.setRptCall2(reflector);
 		m_header.setRptCall1(m_repeaterCall);
+		m_ccsHandler.setReflector(reflector);
 	}
 
 	return res;
@@ -521,7 +640,7 @@ void CDCSGatewayThread::writeLinkingTo(const wxString& callsign)
 			break;
 	}
 
-	m_repeaterHandler->writeText(text, LS_LINKING_DEXTRA, callsign);
+	m_repeaterHandler->writeText(text, LS_LINKING_DCS, callsign);
 	m_audioUnit->setText(text);
 }
 
@@ -562,7 +681,7 @@ void CDCSGatewayThread::writeLinkedTo(const wxString& callsign)
 			break;
 	}
 
-	m_repeaterHandler->writeText(text, LS_LINKED_DEXTRA, callsign);
+	m_repeaterHandler->writeText(text, LS_LINKED_DCS, callsign);
 	m_audioUnit->setText(text);
 }
